Order completion messages, order cancelling, logging & some javadocs

+++ src/OrderClient/Client.java
@@ -8,7 +8,7 @@ public interface Client {
     //Outgoing messages
     int sendOrder(Object par0) throws IOException;
 
-    void sendCancel(long id) throws IOException;
+    void sendCancel(int id) throws IOException;
     //Incoming messages
 
     void partialFill(long order);
